backend/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-whois==0.9.5
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
threadpoolctl==3.6.0
tldextract==5.3.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
